requirements.txt
# Core data
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
imbalanced-learn==0.12.3

# ML models
xgboost==2.0.3
lightgbm==4.3.0

# Deep learning
torch==2.3.0
torchvision==0.18.0

# Explainability
shap==0.45.1

# Visualization & EDA
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0

# Web app
streamlit==1.35.0

# Utilities
pyyaml==6.0.1
joblib==1.4.2
tqdm==4.66.4
python-dotenv==1.0.1

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.29.4
ipywidgets==8.1.3

# Testing
pytest==8.2.1
pytest-cov==5.0.0

# Hyperparameter tuning
optuna==3.6.1

# Data download
kaggle==1.6.14
